Modified the contained letters method to more efficiently parse the wordList for applicable words. Freed up the Server class for implementation of Module 8 material. Made the reset button clear the text fields.
